pyproject.toml
[project]
name = "training-assessment-final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=5.2.8",
    "django-apscheduler>=0.7.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "rest-framework-simplejwt>=0.0.2",
]
